requirements.txt
# Backend
Django==4.2.28
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
mysqlclient==2.2.4
Pillow==12.1.1

# AI / NLP
torch==2.6.0
transformers==4.48.0
sentencepiece==0.2.1
protobuf==5.29.6

# Report processing
pdfplumber==0.11.0
pandas==2.2.2
pytesseract==0.3.10
PyMuPDF==1.24.3
openpyxl==3.1.2

# Frontend
streamlit==1.34.0
requests==2.31.0

# Utilities
python-dotenv==1.0.1
